Add useWebViewHighlight hook to handle postLoadScript
